perf(core): size the voxel record to the stack's real channel count
SubcubeVoxel carried seven per-channel arrays, each dimensioned
MAX_CHANNELS. That provisioned eight channels on every voxel of every
stack: 1376 of its 1436 bytes were per-channel payload, so an L-only
stack used one channel and paid for eight. At 24.5 MP the cube wanted
35.2 GB and ran on swap, measured at 30.2 us/voxel against 9.6 us/voxel
in RAM -- three times slower while fitting five times fewer samples.

Group the per-channel fields into VoxelChannel and store exactly
n_channels of them immediately after the voxel's fixed fields, in one
allocation per cube. The record stays one contiguous struct per (x, y)
with no parallel arrays; only its length becomes a run-time value.
Copying a voxel by value would now silently drop its channels, so the
copy operations are deleted and that mistake is a compile error.

Two lossless packing changes ride along, since both shrink the same
per-channel record:

  - PixelHistogram bins go from uint32 to uint16. One bin is incremented
    per sample and a channel receives at most one sample per frame, so
    no bin can exceed the voxel's frame count -- which is already
    uint16_t. A static_assert ties the two widths together so the proof
    cannot silently lapse. 72 bytes -> 40.

  - ZDistribution declares its two one-byte fields adjacently so they
    share a slot instead of costing four bytes each plus padding.
    64 bytes -> 60. No behaviour change whatsoever.

Per-channel record: 172 bytes -> 136. Per voxel on a 3-channel OSC
stack: 1436 -> 468.

  case                       before    after   saving
  NGC7635 L-only 8.3 MP      11.9 GB   1.6 GB    86%
  M27 LRGB-mono 8.3 MP       11.9 GB   5.0 GB    58%
  M16 / M27 OSC 24.5 MP      35.2 GB  11.5 GB    67%
  ASI6200MC 62 MP            89.0 GB  29.0 GB    67%

The 24.5 MP corpora now fit in RAM on this box instead of thrashing.

The cube also reports its footprint before Phase A, so the number that
decides whether a run fits is in the Process Console rather than in
somebody's head.

Bit-exactness verified, not assumed: the frozen lrgb_mono_ngc7635 E2E
golden still reports stacked c2277834, noise b9ec9edd and all three
stretch sweeps unchanged. 70/70 unit tests green.

+++ src/lib/core/include/nukex/core/histogram.hpp
@@ -17,7 +17,12 @@ namespace nukex {
 struct PixelHistogram {
     static constexpr int N_BINS = 16;
 
-    uint32_t bins[N_BINS] = {};
+    /// One bin is incremented per sample, and a channel receives at most one
+    /// sample per frame, so no bin can exceed the voxel's frame count — which
+    /// is itself uint16_t. 16-bit bins are exactly lossless for every stack
+    /// this codebase can represent, and halve the histogram from 72 to 40
+    /// bytes per channel. test_voxel.cpp static-asserts the two widths match.
+    uint16_t bins[N_BINS] = {};
     float    range_min    = 0.0f;
     float    range_max    = 1.0f;
